laravel/public/rms_script/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.1.0
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Cors==4.0.1
httplib2==0.22.0
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.1.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
prov==2.0.1
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
traits==6.3.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
